Implement equals and hashCode in models
**Summary**
Implement `equals()` and `hashCode()` for all models,
specifically the subclasses of `StripeJsonModel`

**Motivation**
The implementation of `equals()`, `hashCode()`, and
`toString()` in `StripeJsonModel`, which is the base
class of almost all model classes, is extremely
inefficient. It implements these methods by first
transforming the model object to a JSON string, which
is an expensive operation.

**Testing**
- Updated unit tests
- Tested on device with example app